backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai-whisper==20231117
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
pandas==2.1.3

